Stick with the basics: use ALSA and Bluez-ALSA
Pulseaudio is pretty bad. PipeWire and WirePlumber are not sufficient
stable to be taken, really! All this huge solutions are really buggy
and not general to connect with large amount of bluetooth devices.
They have a bunch of problems when switching over different sinks,
requiring reboots and more reboots. It is really sad that all
those projects do not give focus on real general solutions being in
the end only a messy hell of conf patchings following more and more
conf patchings.

Enough!

Let's stick with the basics.