Fix exception when trying to parse extended id3 v2.3
Some songs have the size field formatted in a weird
way: they seem to have 8 bytes (first four zeroed) to
denote the size. From now on we will try to get the
first four, it being zeroed we try to get the next
four. If it is still zeroed we abort the parsing
process.